+++ src/app/components/space-automation/atm-workflow-executions-table/workflow-row.js
@@ -9,8 +9,11 @@
 
 import Component from '@ember/component';
 import { computed } from '@ember/object';
+import { collect } from '@ember/object/computed';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
 import { inject as service } from '@ember/service';
+import { scheduleOnce } from '@ember/runloop';
+import isDirectlyClicked from 'onedata-gui-common/utils/is-directly-clicked';
 
 export default Component.extend(I18n, {
   tagName: 'tr',
@@ -18,6 +21,7 @@ export default Component.extend(I18n, {
   attributeBindings: ['atmWorkflowExecutionSummary.entityId:data-row-id'],
 
   i18n: service(),
+  workflowManager: service(),
 
   /**
    * @override
@@ -45,6 +49,23 @@ export default Component.extend(I18n, {
   onSelect: undefined,
 
   /**
+   * @virtual
+   * @type {Function}
+   * @returns {any}
+   */
+  onCancel: undefined,
+
+  /**
+   * @type {Boolean}
+   */
+  wasCancelledByUser: false,
+
+  /**
+   * @type {Boolean}
+   */
+  areActionsOpened: false,
+
+  /**
    * @type {ComputedProperty<String>}
    */
   statusIcon: computed('atmWorkflowExecutionSummary.status', function statusIcon() {
@@ -55,6 +76,11 @@ export default Component.extend(I18n, {
         return 'time';
       case 'active':
         return 'update';
+      case 'cancelling':
+      case 'cancelled':
+        return 'cancelled';
+      case 'skipped':
+        return 'skipped';
       case 'finished':
         return 'checkbox-filled';
       case 'failed':
@@ -75,9 +101,42 @@ export default Component.extend(I18n, {
     }
   ),
 
-  click() {
+  /**
+   * @type {ComputedProperty<Utils.Action>}
+   */
+  cancelAction: computed(
+    'atmWorkflowExecutionSummary.status',
+    'wasCancelledByUser',
+    function cancelAction() {
+      const status = this.get('atmWorkflowExecutionSummary.status');
+      const wasCancelledByUser = this.get('wasCancelledByUser');
+      const disabled = wasCancelledByUser || [
+        'cancelling',
+        'cancelled',
+        'skipped',
+        'failed',
+        'finished',
+      ].includes(status);
+      return {
+        title: this.t('cancel'),
+        class: 'cancel-atm-workflow-execution-action-trigger',
+        icon: 'cancelled',
+        disabled,
+        action: () => this.cancelAtmWorkflowExecution(),
+      };
+    }
+  ),
+
+  /**
+   * @type {ComputedProperty<Array<Utils.Action>>}
+   */
+  atmWorkflowExecutionActions: collect('cancelAction'),
+
+  click(event) {
     this._super(...arguments);
-    this.notifyAboutSelection();
+    if (isDirectlyClicked(event, this.get('element'))) {
+      this.notifyAboutSelection();
+    }
   },
 
   notifyAboutSelection() {
@@ -86,5 +145,22 @@ export default Component.extend(I18n, {
       onSelect,
     } = this.getProperties('atmWorkflowExecutionSummary', 'onSelect');
     onSelect && onSelect(atmWorkflowExecutionSummary);
+  },
+
+  async cancelAtmWorkflowExecution() {
+    this.set('wasCancelledByUser', true);
+    const {
+      workflowManager,
+      onCancel,
+    } = this.getProperties('workflowManager', 'onCancel');
+    const atmWorkflowExecutionId = this.get('atmWorkflowExecutionSummary.entityId');
+    await workflowManager.cancelAtmWorkflowExecution(atmWorkflowExecutionId);
+    onCancel && onCancel();
+  },
+
+  actions: {
+    toggleActionsOpen(state) {
+      scheduleOnce('afterRender', this, 'set', 'areActionsOpened', state);
+    },
   },
 });