+++ src/app/locales/en/components/space-automation/atm-workflow-executions-table/workflow-row.js
@@ -5,7 +5,11 @@ export default {
     preparing: 'Preparing',
     enqueued: 'Enqueued',
     active: 'Active',
+    cancelling: 'Cancelling',
+    cancelled: 'Cancelled',
+    skipped: 'Skipped',
     finished: 'Finished',
     failed: 'Failed',
   },
+  cancel: 'Cancel',
 };